Merges up through r18894 from the original trunk
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/acstools/trunk@18913 fe389314-cf27-0410-b35b-8c050e845b92

+++ acstools/runastrodriz.py
@@ -1,27 +1,45 @@
 #!/usr/bin/env python
 
-""" runastrodriz.py - Module to control operation of astrodrizzle to 
+""" runastrodriz.py - Module to control operation of astrodrizzle to
         remove distortion and combine HST images in the pipeline.
 
-USAGE: runastrodriz.py inputFilename 
+USAGE: runastrodriz.py [-fhi] inputFilename [newpath]
 
-Alternative USAGE: 
+Alternative USAGE:
     python
     from acstools import runastrodriz
-    runastrodriz.process(inputFilename,force=False)
+    runastrodriz.process(inputFilename,force=False,newpath=None,inmemory=False)
 
 GUI Usage under Python:
     python
     from stsci.tools import teal
     import acstools
     cfg = teal.teal('runastrodriz')
-    
+
 PyRAF Usage:
     epar runastrodriz
-    
+
+If the '-i' option gets specified, no intermediate products will be written out
+to disk. These products, instead, will be kept in memory. This includes all
+single drizzle products (*single_sci and *single_wht), median image,
+blot images, and crmask images.  The use of this option will therefore require
+significantly more memory than usual to process the data.
+
+If a value has been provided for the newpath parameter, all processing will be
+performed in that directory/ramdisk.  The steps involved are:
+   - create a temporary directory under that directory named after the input file
+   - copy all files related to the input to that new directory
+   - change to that new directory and run astrodrizzle
+   - change back to original directory
+   - move (not copy) ALL files from temp directory to original directory
+   - delete temp sub-directory
+
 *** INITIAL VERSION
-W.J. Hack  12 Aug 2011: Initial version based on Version 1.2.0 of 
+W.J. Hack  12 Aug 2011: Initial version based on Version 1.2.0 of
                         STSDAS$pkg/hst_calib/wfc3/runwf3driz.py
+W.J. Hack  27 Jun 2012: Implement support to process in different directory
+
+W.J. Hack  24 Aug 2012: Provided interface for in-memory option
 """
 
 # Import standard Python modules
@@ -36,11 +54,11 @@ import pyfits
 
 __taskname__ = "runastrodriz"
 
-# Local variables 
-__version__ = "1.1.1"
-__vdate__ = "(14-Sep-2011)"
+# Local variables
+__version__ = "1.3.0"
+__vdate__ = "(24-Aug-2012)"
 
-# Define parameters which need to be set specifically for 
+# Define parameters which need to be set specifically for
 #    pipeline use of astrodrizzle
 pipeline_pars = {'mdriztab':True,
                  'stepsize':10,
@@ -48,13 +66,13 @@ pipeline_pars = {'mdriztab':True,
                  'updatewcs':True,
                  'preserve':False,
                  'resetbits':4096}
-                
+
 #default marker for trailer files
 __trlmarker__ = '*** astrodrizzle Processing Version '+__version__+__vdate__+'***\n'
 
 # History:
 # Version 1.0.0 - Derived from v1.2.0 of wfc3.runwf3driz to run astrodrizzle
-    
+
 #### TEAL Interfaces
 def getHelpAsString():
     helpString = 'runastrodriz Version '+__version__+__vdate__+'\n'
@@ -66,16 +84,17 @@ def help():
     print getHelpAsString()
 
 def run(configobj=None):
-    process(configobj['input'],force=configobj['force'])
-       
+    process(configobj['input'],force=configobj['force'],
+                newpath=configobj['newpath'],inmemory=configobj['in_memory'])
+
 #### Primary user interface
-def process(inFile,force=False):
-    """ Run astrodrizzle on input file/ASN table 
+def process(inFile,force=False,newpath=None, inmemory=False):
+    """ Run astrodrizzle on input file/ASN table
         using default values for astrodrizzle parameters.
     """
     # We only need to import this package if a user run the task
-    import astrodither
-    from astrodither import processInput # used for creating new ASNs for _flc inputs
+    import drizzlepac
+    from drizzlepac import processInput # used for creating new ASNs for _flc inputs
 
     # Open the input file
     try:
@@ -87,18 +106,27 @@ def process(inFile,force=False):
     except TypeError:
         print "ERROR: Inappropriate input file."
         return
-        
+
+    #If newpath was specified, move all files to that directory for processing
+    if newpath:
+        orig_processing_dir = os.getcwd()
+        new_processing_dir = _createWorkingDir(newpath,inFilename)
+        _copyToNewWorkingDir(new_processing_dir,inFilename)
+        os.chdir(new_processing_dir)
+
     # Initialize for later use...
     _mname = None
+    _new_asn = None
     # Check input file to see if [DRIZ/DITH]CORR is set to PERFORM
     if '_asn' in inFilename:
         # We are working with an ASN table.
         # Use asnutil code to extract filename
         inFilename = _lowerAsn(inFilename)
+        _new_asn = [inFilename]
         _asndict = asnutil.readASNTable(inFilename,None,prodonly=False)
         _fname = fileutil.buildRootname(string.lower(_asndict['output']),ext=['_drz.fits'])
         _cal_prodname = string.lower(_asndict['output'])
-        
+
         # Retrieve the first member's rootname for possible use later
         _fimg = pyfits.open(inFilename)
         for name in _fimg[1].data.field('MEMNAME'):
@@ -107,7 +135,7 @@ def process(inFile,force=False):
                 break
         _fimg.close()
         del _fimg
-    
+
     else:
         # Check to see if input is a _RAW file
         # If it is, strip off the _raw.fits extension...
@@ -118,14 +146,14 @@ def process(inFile,force=False):
         _cal_prodname = inFilename[:_indx]
         # Reset inFilename to correspond to appropriate input for
         # drizzle: calibrated product name.
-        inFilename = _mname        
+        inFilename = _mname
 
         if _mname == None:
-            errorMsg = 'Could not find calibrated product!' 
+            errorMsg = 'Could not find calibrated product!'
             raise Exception,errorMsg
 
     # Create trailer filenames based on ASN output filename or
-    # on input name for single exposures    
+    # on input name for single exposures
     if '_raw' in inFile:
         # Output trailer file to RAW file's trailer
         _trlroot = inFile[:inFile.find('_raw')]
@@ -145,7 +173,7 @@ def process(inFile,force=False):
 
     # Open product and read keyword value
     # Check to see if product already exists...
-    dkey = 'DRIZCORR'  
+    dkey = 'DRIZCORR'
     # ...if product does NOT exist, interrogate input file
     # to find out whether 'dcorr' has been set to PERFORM
     # Check if user wants to process again regardless of DRIZCORR keyword value
@@ -162,14 +190,14 @@ def process(inFile,force=False):
             del _fimg
         else:
             dcorr = None
-    
+
     time_str = _getTime()
     _tmptrl = _trlroot + '_tmp.tra'
     _drizfile = _trlroot + '_pydriz'
     _drizlog = _drizfile + ".log" # the '.log' gets added automatically by astrodrizzle
-    
-    _new_asn = None
-    if dcorr == 'PERFORM':    
+
+    print 'dcorr = ',dcorr,inFilename
+    if dcorr == 'PERFORM':
         if '_asn.fits' not in inFilename:
             # Working with a singleton
             # However, we always want to make sure we always use
@@ -180,33 +208,33 @@ def process(inFile,force=False):
             _cal_prodname = _infile
             _inlist = [_infile]
             # Add CTE corrected filename as additional input if present
-            if _infile_flc != _infile:
+            if os.path.exists(_infile_flc) and _infile_flc != _infile:
                 _inlist.append(_infile_flc)
 
         else:
             # Working with an ASN table...
             _infile = inFilename
-            flist,duplist = processInput.checkForDuplicateInputs(_asndict)
+            flist,duplist = processInput.checkForDuplicateInputs(_asndict['order'])
             if len(duplist) > 0:
                 origasn = processInput.changeSuffixinASN(inFilename,'flt')
                 dupasn = processInput.changeSuffixinASN(inFilename,'flc')
                 _inlist = [origasn,dupasn]
             else:
                 _inlist = [_infile]
-            # We want to keep the original specification of the calibration 
+            # We want to keep the original specification of the calibration
             # product name, though, not a lower-case version...
             _cal_prodname = inFilename
-            _new_asn = _inlist # kept so we can delete it when finished
-            
-        
+            _new_asn.extend(_inlist) # kept so we can delete it when finished
+
+
         # Run astrodrizzle and send its processing statements to _trlfile
-        _pyver = astrodither.__version__
-        
+        _pyver = drizzlepac.astrodrizzle.__version__
+
         for _infile in _inlist: # Run astrodrizzle for all inputs
             # Create trailer marker message for start of astrodrizzle processing
             _trlmsg = _timestamp('astrodrizzle started ')
             _trlmsg = _trlmsg+ __trlmarker__
-            _trlmsg = _trlmsg + '%s: Processing %s with astrodrizzle Version %s\n' % (time_str,_infile,_pyver)   
+            _trlmsg = _trlmsg + '%s: Processing %s with astrodrizzle Version %s\n' % (time_str,_infile,_pyver)
             print _trlmsg
 
             # Write out trailer comments to trailer file...
@@ -214,28 +242,29 @@ def process(inFile,force=False):
             ftmp.writelines(_trlmsg)
             ftmp.close()
             _appendTrlFile(_trlfile,_tmptrl)
-            
+
             _pyd_err = _trlroot+'_pydriz.stderr'
 
             try:
-                b = astrodither.AstroDrizzle(input=_infile,runfile=_drizfile,
-                                            configObj='defaults',**pipeline_pars)
+                b = drizzlepac.astrodrizzle.AstroDrizzle(input=_infile,runfile=_drizfile,
+                                            configobj='defaults',in_memory=inmemory,
+                                            **pipeline_pars)
             except Exception, errorobj:
                 _appendTrlFile(_trlfile,_drizlog)
                 _appendTrlFile(_trlfile,_pyd_err)
                 _ftrl = open(_trlfile,'a')
-                _ftrl.write('ERROR: Could not complete astrodrizzle processing of %s.\n' % _infile) 
+                _ftrl.write('ERROR: Could not complete astrodrizzle processing of %s.\n' % _infile)
                 _ftrl.write(str(sys.exc_type)+': ')
                 _ftrl.writelines(str(errorobj))
                 _ftrl.write('\n')
                 _ftrl.close()
                 print 'ERROR: Could not complete astrodrizzle processing of %s.' % _infile
                 raise Exception, str(errorobj)
-            
+
             # Now, append comments created by PyDrizzle to CALXXX trailer file
             print 'Updating trailer file %s with astrodrizzle comments.' % _trlfile
             _appendTrlFile(_trlfile,_drizlog)
-        
+
         # Save this for when PyFITS can modify a file 'in-place'
         # Update calibration switch
         _fimg = pyfits.open(_cal_prodname,mode='update')
@@ -249,7 +278,7 @@ def process(inFile,force=False):
         for _prodname in _prodlist:
             _plower = _prodname.lower()
             if _prodname != _plower:  os.rename(_prodname,_plower)
-                
+
     else:
         # Create default trailer file messages when astrodrizzle is not
         # run on a file.  This will typically apply only to BIAS,DARK
@@ -260,31 +289,43 @@ def process(inFile,force=False):
         _trlmsg = _trlmsg + '%s: astrodrizzle processing not requested for %s.\n' % (time_str,inFilename)
         _trlmsg = _trlmsg + '       astrodrizzle will not be run at this time.\n'
         print _trlmsg
-        
+
         # Write message out to temp file and append it to full trailer file
         ftmp = open(_tmptrl,'w')
         ftmp.writelines(_trlmsg)
         ftmp.close()
         _appendTrlFile(_trlfile,_tmptrl)
-    
-    _fmsg = None    
+
+    _fmsg = None
     # Append final timestamp to trailer file...
-    _final_msg = '%s: Finished processing %s \n' % (time_str,inFilename)   
+    _final_msg = '%s: Finished processing %s \n' % (time_str,inFilename)
     _final_msg += _timestamp('astrodrizzle completed ')
     _trlmsg += _final_msg
     ftmp = open(_tmptrl,'w')
     ftmp.writelines(_trlmsg)
     ftmp.close()
     _appendTrlFile(_trlfile,_tmptrl)
-    
+
     # If we created a new ASN table, we need to remove it
     if _new_asn != None:
-        for _name in _new_asn: os.remove(_name)
-        
+        for _name in _new_asn: fileutil.removeFile(_name)
+
     # Clean up any generated OrIg_files directory
     if os.path.exists("OrIg_files"):
-        os.rmdir("OrIg_files")
-    
+        # check to see whether this directory is empty
+        flist = glob.glob('OrIg_files/*.fits')
+        if len(flist) == 0:
+            os.rmdir("OrIg_files")
+        else:
+            print 'OrIg_files directory NOT removed as it still contained images...'
+
+    # If processing was done in a temp working dir, restore results to original
+    # processing directory, return to original working dir and remove temp dir
+    if newpath:
+        _restoreResults(new_processing_dir,orig_processing_dir)
+        os.chdir(orig_processing_dir)
+        _removeWorkingDir(new_processing_dir)
+
     # Provide feedback to user
     print _final_msg
 
@@ -297,17 +338,17 @@ def _lowerAsn(asnfile):
     _new_asn = asnfile[:_indx]+'_pipeline'+asnfile[_indx:]
     if os.path.exists(_new_asn) == True:
         os.remove(_new_asn)
-    # copy original ASN table to new table 
+    # copy original ASN table to new table
     shutil.copy(asnfile,_new_asn)
-    
+
     # Open up the new copy and convert all MEMNAME's to lower-case
     fasn = pyfits.open(_new_asn,'update')
     for i in xrange(len(fasn[1].data)):
         fasn[1].data[i].setfield('MEMNAME',fasn[1].data[i].field('MEMNAME').lower())
     fasn.close()
-    
+
     return _new_asn
-    
+
 def _appendTrlFile(trlfile,drizfile):
     """ Append drizfile to already existing trlfile from CALXXX.
     """
@@ -317,17 +358,17 @@ def _appendTrlFile(trlfile,drizfile):
     ftrl = open(trlfile,'a')
     # Open astrodrizzle trailer file
     fdriz = open(drizfile)
-    
+
     # Read in drizzle comments
     _dlines = fdriz.readlines()
-    
+
     # Append them to CALWF3 trailer file
     ftrl.writelines(_dlines)
-    
+
     # Close all files
     ftrl.close()
     fdriz.close()
-    
+
     # Now, clean up astrodrizzle trailer file
     os.remove(drizfile)
 
@@ -337,13 +378,56 @@ def _timestamp(_process_name):
     _prefix= time.strftime("%Y%j%H%M%S-I-----",time.localtime())
     _lenstr = 60 - len(_process_name)
     return _prefix+_process_name+(_lenstr*'-')+'\n'
-    
+
 def _getTime():
-    # Format time values for keywords IRAF-TLM, and DATE 
+    # Format time values for keywords IRAF-TLM, and DATE
     _ltime = time.localtime(time.time())
     time_str = time.strftime('%H:%M:%S (%d-%b-%Y)',_ltime)
-    
+
     return time_str
+
+#### Functions used to manage processing in a separate directory/ramdisk
+def _createWorkingDir(rootdir,input):
+    """
+    Create a working directory based on input name under the parent directory specified as rootdir
+    """
+    # extract rootname from input
+    rootname = input[:input.find('_')]
+    newdir = os.path.join(rootdir,rootname)
+    if not os.path.exists(newdir):
+        os.mkdir(newdir)
+    return newdir
+
+def _copyToNewWorkingDir(newdir,input):
+    """ Copy input file and all related files necessary for processing to the new working directory.
+
+        This function works in a greedy manner, in that all files associated
+        with all inputs(have the same rootname) will be copied to the new
+        working directory.
+    """
+    flist = []
+    if '_asn.fits' in input:
+        asndict = asnutil.readASNTable(input,None)
+        flist.append(input[:input.find('_')])
+        flist.extend(asndict['order'])
+        flist.append(asndict['output'])
+    else:
+        flist.append(input[:input.find('_')])
+    # copy all files related to these rootnames into new dir
+    for rootname in flist:
+        for fname in glob.glob(rootname+'*'):
+            shutil.copy(fname,os.path.join(newdir,fname))
+
+def _restoreResults(newdir,origdir):
+    """ Move (not copy) all files from newdir back to the original directory
+    """
+    for fname in glob.glob(os.path.join(newdir,'*')):
+        shutil.move(fname,os.path.join(origdir,os.path.basename(fname)))
+
+def _removeWorkingDir(newdir):
+    """ Delete working directory
+    """
+    os.rmdir(newdir)
 
 #### Functions to support execution from the shell.
 def main():
@@ -351,7 +435,7 @@ def main():
     import getopt
 
     try:
-        optlist, args = getopt.getopt(sys.argv[1:], 'hf')
+        optlist, args = getopt.getopt(sys.argv[1:], 'hfi')
     except getopt.error, e:
         print str(e)
         print __doc__
@@ -360,7 +444,9 @@ def main():
     # initialize default values
     help = 0
     force = False
-    
+    newdir = None
+    inmemory = False
+
     # read options
     for opt, value in optlist:
         if opt == "-h":
@@ -368,22 +454,25 @@ def main():
         if opt == "-f":
             force = True
         if opt == "-i":
-            interactive = True
+            inmemory = True
+
     if len(args) < 1:
-        print "syntax: runastrodriz.py [-fh] inputFilename"
+        print "syntax: runastrodriz.py [-fhi] inputFilename [newpath]"
         sys.exit()
+    if len(args) > 1:
+        newdir = args[-1]
     if (help):
         print __doc__
         print "\t", __version__+'('+__vdate__+')'
     else:
         try:
-            process(args[0],force=force)
+            process(args[0],force=force,newpath=newdir, inmemory=inmemory)
         except Exception, errorobj:
             print str(errorobj)
             print "ERROR: Cannot run astrodrizzle on %s." % sys.argv[1]
             raise Exception, str(errorobj)
-        
+
     sys.exit()
 
 if __name__ == "__main__":
-    main() 
+    main()
